requirements.txt
pygame
opencv-python
numpy
pyyaml
pyserial